GH-80 - Add Spring Data REST example.
This also includes setting the mapping context for Spring internals.
The example can also be used as a dev container with vs code.

+++ examples/multi-database/src/main/java/org/neo4j/springframework/data/examples/spring_boot/domain/PersonEntity.java
@@ -30,14 +30,14 @@ public class PersonEntity {
 	@Id
 	private final String name;
 
-	private final Long born;
+	private final Integer born;
 
-	public PersonEntity(Long born, String name) {
+	public PersonEntity(Integer born, String name) {
 		this.born = born;
 		this.name = name;
 	}
 
-	public Long getBorn() {
+	public Integer getBorn() {
 		return born;
 	}
